refactor(3139): lift SSTableRowIteratorAdapter + path-based producer thread out of merge/mod.rs
Pure code motion (epic #1116 campsite rule): merge/mod.rs was 13,558 lines.
Moves the adapter struct, RECV_CANCEL_POLL, open/producer_thread, the
SSTableRowIterator consumer impl and the cancel-aware Drop teardown into
merge/producer_iter.rs, and the reader->MergeEntry conversion helpers into
merge/producer_iter_convert.rs. Moved lines are byte-identical apart from
use paths, pub(super) visibility the move requires, and three intra-doc
links requalified with super::. No logic, signature or gauge-accounting
change.
