+++ cqlite-core/src/storage/write_engine/merge/mod.rs
@@ -421,107 +421,24 @@ mod async_bridge;
 #[cfg(feature = "write-support")]
 pub(crate) use async_bridge::block_on_async;
 
-/// Adapter that wraps async SSTableReader into a true-streaming sync
-/// [`SSTableRowIterator`].
-///
-/// ## Design (Issue #754 — remove 128MB buffer cap residue of #447)
-///
-/// The V5CompressedLegacy format requires chunk stitching: a partition may
-/// straddle compression-chunk boundaries, so the decoder needs a contiguous
-/// view spanning at least one whole partition. The reader's streaming path
-/// keeps only a **sliding window** of that view — one chunk plus the partition
-/// currently being decoded — rather than the whole decompressed file.
-///
-/// A background thread (the producer) opens the SSTable with its own Tokio
-/// runtime and calls
-/// [`stream_all_partitions_for_compaction`](crate::storage::sstable::reader::SSTableReader::stream_all_partitions_for_compaction),
-/// which decompresses one chunk at a time, drains every fully-decoded partition
-/// out of the window, and forwards each entry one at a time into a bounded
-/// `sync_channel`. The channel capacity is up to [`STREAMING_CHANNEL_CAPACITY`]
-/// entries, adaptively reduced under concurrent merges (see [`egress_budget`]);
-/// once the channel is full the producer blocks until the consumer (the main
-/// merge thread) pulls the next entry.
-///
-/// The bounded window plus the bounded channel together make end-to-end peak
-/// memory independent of total input size: a source's decompressed content is
-/// never fully resident. Peak is roughly `max_partition_size + one_chunk +
-/// channel_capacity` per source (issue #827).
-///
-/// ## Issue #591 safety (mmap vs file deletion)
-///
-/// `finalize_merge_async` deletes the input SSTable files once the merged output
-/// is published. We require that no mmap outlives its backing file. The producer
-/// thread opens the reader with `use_mmap = false`, and the thread *fully reads
-/// all file data* (the stitching phase) before it can block on a channel send.
-/// By the time `finalize_merge_async` runs, the merge is complete and all
-/// channel entries have been consumed, so the producer thread has long since
-/// finished and dropped its file handle. No mmap ever exists.
-///
-/// ## Issue #587 safety (async-from-sync bridge)
-///
-/// The producer thread creates its own Tokio runtime (never `Handle::block_on`),
-/// so it cannot panic even when called from within an existing Tokio runtime.
-/// This is the same strategy as [`block_on_async`].
-///
-/// ## Issue #2316 thread budget (O(M), no per-producer worker pool)
-///
-/// That per-producer runtime is a **`current_thread`** runtime, NOT a
-/// multi-threaded `Runtime::new()`: it drives the producer's single sequential
-/// scan (no internal `tokio::spawn`) on the producer thread itself, adding ZERO
-/// worker threads. A merge over `M` inputs therefore costs `O(M)` OS threads, not
-/// `M + M·num_cpus` — killing the context-switch storm under concurrent `do_get`.
+/// Path-based streaming producer-thread row iterator (issue #3139): the
+/// [`SSTableRowIteratorAdapter`] that wraps an async `SSTableReader` into a sync
+/// [`SSTableRowIterator`], its producer thread, the consumer side, and the
+/// cancel-aware join-on-drop teardown. Lifted VERBATIM out of this file per the
+/// #1116 campsite rule (behaviour unchanged); the shared-reader producer shape
+/// stays in [`from_readers`].
 #[cfg(feature = "write-support")]
-struct SSTableRowIteratorAdapter {
-    /// Receiving end of the bounded channel fed by the producer thread. `Option`
-    /// so [`Drop`] can drop it FIRST (issue #2361) — closing the channel wakes a
-    /// producer blocked on a full `SyncSender::send` (its send returns `Err`), so
-    /// the subsequent producer join is bounded and cannot deadlock.
-    receiver:
-        Option<std::sync::mpsc::Receiver<std::result::Result<MergeEntry, MergeProducerError>>>,
-    /// Producer thread handle. `Option` so [`Drop`] can `take()` it and JOIN the
-    /// thread (issue #2361) rather than detach it — the pre-#2361 `_producer`
-    /// field's "joined on drop" doc was WRONG (dropping a `JoinHandle` detaches
-    /// the thread, leaving it to run to completion in the background; a cancelled
-    /// `do_get` over a 1M-partition scan would keep burning CPU/IO). Joining after
-    /// dropping the receiver + tripping `scan_cancel` guarantees prompt teardown.
-    producer: Option<std::thread::JoinHandle<()>>,
-    /// The PER-CALL cancellation token this adapter's producer scan polls (issues
-    /// #2264/#2346). Held so (a) [`Self::next`] can make its blocking channel
-    /// `recv` cancel-aware and (b) [`Drop`] can trip it, waking a producer that is
-    /// mid-scan (not blocked on send) so it exits promptly before the join.
-    scan_cancel: crate::storage::scan_cancel::ScanCancel,
-    /// This adapter's OWN count of DATA entries its producer thread has
-    /// successfully sent into the bounded channel (issue #2419 roborev job
-    /// 1733), shared with the producer thread via `Arc` so `Drop` can read it.
-    /// It only becomes STABLE — no further increments possible — once the
-    /// producer thread has been joined (see `Drop`). Paired with
-    /// [`Self::received_count`] to compute the exact post-join egress-depth
-    /// reconcile residual: reading it BEFORE the join (as the pre-fix drain loop
-    /// did) races a producer send that can slip in after the drain's last
-    /// `Empty` and before the receiver is actually dropped, permanently leaking
-    /// the shared `cqlite.merge.egress_channel_depth` gauge upward across
-    /// repeated cancellations on a long-running server.
-    sent_count: std::sync::Arc<std::sync::atomic::AtomicI64>,
-    /// This adapter's own count of DATA entries actually received/consumed,
-    /// incremented alongside every [`channel_depth::received`] call this adapter
-    /// makes ([`Self::next`]'s normal consumption). Only ever touched by the
-    /// thread holding `&mut self` (never shared), so a plain `i64`.
-    received_count: i64,
-    /// Test-only (issue #2765): the exact `sync_channel` capacity this adapter's
-    /// egress channel was built with — the merge-scoped adaptive snapshot the
-    /// constructor threaded in. Observed via [`SSTableRowIterator::egress_channel_capacity`]
-    /// so a wiring test proves the budget reaches BOTH construction sites.
-    #[cfg(test)]
-    egress_channel_capacity: usize,
-}
+mod producer_iter;
+#[cfg(feature = "write-support")]
+use producer_iter::SSTableRowIteratorAdapter;
 
-/// Poll interval for the cancel-aware blocking `recv` in
-/// [`SSTableRowIteratorAdapter::next`] (issue #2361). A bounded wait so a
-/// cancellation that lands while the consumer is blocked waiting for the next
-/// producer entry is observed within one interval, without a busy-spin. This is
-/// an INTERNAL cadence only — never asserted against wall-clock in tests.
+/// Reader→merge row/cell conversion helpers for [`SSTableRowIteratorAdapter`]
+/// (issue #3139): `CompactionRow` → [`MergeEntry`] / [`RowData`] translation,
+/// clustering-key extraction and range-bound translation, shared by BOTH
+/// producer shapes. An inherent `impl` block in a sibling file (as
+/// [`from_readers`] does), lifted verbatim out of this file per #1116.
 #[cfg(feature = "write-support")]
-const RECV_CANCEL_POLL: std::time::Duration = std::time::Duration::from_millis(50);
+mod producer_iter_convert;
 
 /// Channel-safe error payload sent from a producer thread to the merge (issue
 /// #2264).
@@ -612,813 +529,6 @@ mod reconcile_microalloc_tests;
 // linear scan (sibling file, #1116 campsite rule).
 #[cfg(all(test, feature = "write-support"))]
 mod range_shadowing_binsearch_tests;
-
-#[cfg(feature = "write-support")]
-impl SSTableRowIteratorAdapter {
-    /// Open an SSTable and start a streaming producer thread.
-    ///
-    /// Returns immediately; the producer thread runs concurrently and populates
-    /// the channel as the consumer advances. The file handle is held only by
-    /// the producer thread and is dropped when the thread finishes.
-    ///
-    /// Uses [`SSTableReader::iterate_all_partitions_for_compaction`] which
-    /// returns actual per-row timestamps decoded from the on-disk row headers.
-    /// This allows the k-way merger to perform timestamp-accurate last-write-wins
-    /// ordering, which is essential for tombstone shadowing (Issue #505).
-    ///
-    /// When the schema has clustering columns, their values are extracted from
-    /// the decoded cells (in the producer thread, by column name in schema order)
-    /// and stored on `MergeEntry.clustering_key` so `merge_partition_rows` groups
-    /// and reconciles distinct clustering rows correctly. The clustering columns
-    /// are left in the cells as well, since the read-back path expects them there.
-    ///
-    /// `channel_capacity` (issue #2765) is the merge-scoped adaptive egress
-    /// capacity snapshotted ONCE by the `KWayMerger` constructor and shared by
-    /// every source channel of that merge — never derived per adapter here.
-    fn open(
-        path: &Path,
-        run_index: usize,
-        schema: &TableSchema,
-        udt_registry: Option<crate::schema::UdtRegistry>,
-        scan_cancel: crate::storage::scan_cancel::ScanCancel,
-        channel_capacity: usize,
-    ) -> Result<Self> {
-        let path_buf = path.to_path_buf();
-        let schema = schema.clone();
-        // Held on the adapter for cancel-aware recv + Drop teardown (issue #2361).
-        let adapter_cancel = scan_cancel.clone();
-
-        let (sender, receiver) = std::sync::mpsc::sync_channel(channel_capacity);
-        // Issue #2419 roborev job 1733: this adapter's own sent-count, shared
-        // with the producer thread so `Drop` can read its post-join-stable value
-        // for the egress-depth reconcile (see the field doc).
-        let sent_count = std::sync::Arc::new(std::sync::atomic::AtomicI64::new(0));
-        let producer_sent_count = sent_count.clone();
-
-        // Issue #2316: account this producer on the `cqlite.merge.producer_threads`
-        // gauge BEFORE spawning, so the increment happens-before any possible
-        // decrement. The decrementing `ProducerThreadGuard` is created first thing
-        // inside the child thread (see `producer_thread`); a fast-exiting producer
-        // (e.g. a reader-open error) could otherwise race its own guard's drop
-        // against a post-spawn increment on the parent, transiently underflowing
-        // the live count. Incrementing here first makes the pairing
-        // correct-by-construction — no ordering race is possible.
-        producer_gauge::spawned();
-
-        // Issue #2819: propagate the sub-phase sink onto this path-based producer
-        // thread too (thread-locals are not inherited across a spawn); `None`
-        // (no-op) for non-flight callers. NOTE (roborev L2): this is a DEFENSIVE /
-        // latent propagation, NOT a parity-covered do_get path — production
-        // `do_get` is warm-only (`spawn_streaming_from_readers` → `open_from_reader`,
-        // which has the tested propagation). The path-based `open`
-        // (`MergeInput::Paths`) is the test-only byte-identity oracle, so its
-        // cold_fault/decompress attribution has no e2e assertion (and would fire
-        // only for a stitching/BTI fixture via the `read_next_block` loop). Kept so
-        // a future path-based flight caller is correct by construction, not to
-        // imply current parity coverage. (This file is far over the #1116 campsite
-        // target; the +2 lines are the minimal propagation.)
-        let subphase_sink = crate::observability::stream_subphase::current();
-
-        // Spawn the producer thread via `Builder::spawn` (rather than the
-        // panic-on-failure `std::thread::spawn`) so an OS thread-creation failure
-        // is a recoverable `Err`, not a process abort: the gauge increment above
-        // must never leak for a thread that never actually started, so roll it
-        // back on this path via `producer_gauge::rollback`. The thread owns a
-        // fresh single-threaded (current_thread) Tokio runtime so it never
-        // collides with any runtime on the calling thread (Issue #587) and adds no
-        // worker threads beyond itself (Issue #2316).
-        let producer = match std::thread::Builder::new().spawn(move || {
-            let _subphase_guard = crate::observability::stream_subphase::install(subphase_sink);
-            Self::producer_thread(
-                path_buf,
-                run_index,
-                schema,
-                udt_registry,
-                scan_cancel,
-                sender,
-                producer_sent_count,
-            );
-        }) {
-            Ok(handle) => handle,
-            Err(e) => {
-                producer_gauge::rollback();
-                return Err(Error::Storage(format!(
-                    "streaming producer: failed to spawn thread: {}",
-                    e
-                )));
-            }
-        };
-
-        Ok(Self {
-            receiver: Some(receiver),
-            producer: Some(producer),
-            scan_cancel: adapter_cancel,
-            sent_count,
-            received_count: 0,
-            #[cfg(test)]
-            egress_channel_capacity: channel_capacity,
-        })
-    }
-
-    /// Body of the producer thread.
-    ///
-    /// Opens the SSTable with buffered I/O (Issue #591), then **streams** the
-    /// source one partition at a time via
-    /// [`stream_all_partitions_for_compaction`](crate::storage::sstable::reader::SSTableReader::stream_all_partitions_for_compaction),
-    /// converting each entry to a [`MergeEntry`] (populating the clustering key
-    /// from the decoded cells when the schema has clustering columns) and
-    /// sending it through the bounded channel immediately (issue #827). The
-    /// blocking `SyncSender::send` provides the backpressure that — together
-    /// with the reader's sliding-window stitch+parse — keeps peak memory bounded
-    /// by `max_partition_size + one_chunk + channel_capacity`, independent of
-    /// the total source size. Errors are forwarded as `Err(String)`.
-    fn producer_thread(
-        path_buf: PathBuf,
-        run_index: usize,
-        schema: TableSchema,
-        udt_registry: Option<crate::schema::UdtRegistry>,
-        scan_cancel: crate::storage::scan_cancel::ScanCancel,
-        sender: std::sync::mpsc::SyncSender<std::result::Result<MergeEntry, MergeProducerError>>,
-        sent_count: std::sync::Arc<std::sync::atomic::AtomicI64>,
-    ) {
-        // Issue #2316: decrement the live producer-thread gauge when this thread
-        // exits (even on panic). Created FIRST so the spawn-time increment in
-        // `open` is always balanced exactly once.
-        let _thread_guard = producer_gauge::ProducerThreadGuard;
-
-        // Drive the streaming read on an owned Tokio runtime (Issue #587): the
-        // producer owns its single-purpose runtime, so the blocking
-        // `SyncSender::send` inside the emit callback never stalls a shared
-        // runtime, and there is no nested `block_on` / `Handle::current`. The
-        // runtime is `current_thread` (Issue #2316), so it adds no worker threads.
-        // Buffered I/O (Issue #591): the file must not be memory-mapped OR read
-        // via direct I/O because finalize_merge_async may delete it after the
-        // merge completes. The default disk-access mode is `Auto`, which would
-        // otherwise map (or direct-read) inputs above the size thresholds, so we
-        // force `Buffered` explicitly here — clearing `use_mmap` alone is no
-        // longer sufficient now that `Auto` ignores that legacy flag.
-        // Clone the sender for the error path: the streaming closure moves one
-        // clone for per-entry sends, leaving this one to report a fatal error.
-        let error_sender = sender.clone();
-        let stream_result = (|| -> Result<()> {
-            use crate::config::DiskAccessMode;
-            use crate::platform::Platform;
-            use crate::Config;
-            use std::sync::Arc;
-
-            let mut config = Config::default();
-            config.storage.use_mmap = false;
-            config.storage.disk_access_mode = DiskAccessMode::Buffered;
-
-            // Issue #2316: a `current_thread` runtime drives the scan ON this
-            // producer thread with ZERO extra workers. A multi-threaded
-            // `Runtime::new()` would spin up `num_cpus` workers per producer
-            // (~M·num_cpus threads/merge) that the single sequential scan never
-            // uses — pure overhead. This bounds the per-merge thread cost to O(M).
-            let rt = tokio::runtime::Builder::new_current_thread()
-                .enable_all()
-                .build()
-                .map_err(|e| {
-                    Error::Storage(format!(
-                        "streaming producer: failed to create runtime: {}",
-                        e
-                    ))
-                })?;
-
-            rt.block_on(async move {
-                let platform = Arc::new(Platform::new(&config).await?);
-                let mut reader = crate::storage::sstable::reader::SSTableReader::open(
-                    &path_buf, &config, platform,
-                )
-                .await?;
-
-                // Issue #1234: wire the authoritative UDT registry onto the reader
-                // so the compaction read path decodes a top-level `frozen<UDT>`
-                // value structurally. Without it the value decode errors
-                // (`UDT not found in registry`), the row reconciles to empty, and
-                // the partition is dropped — silent data loss during compaction.
-                // This mutation requires `&mut self`, so it can only happen HERE
-                // — before the reader is ever shared — not on the shared-reader
-                // producer path (issue #2346, see `from_readers::open_from_reader`'s
-                // doc comment for that seam's UDT-registry contract).
-                if let Some(registry) = udt_registry {
-                    reader.set_udt_registry(registry);
-                }
-
-                // Issue #2264/#2346: the cooperative-cancellation token is now a
-                // PER-CALL parameter to `stream_all_partitions_for_compaction`
-                // rather than mutated onto the reader (`set_scan_cancel` needed
-                // `&mut self`, which a SHARED reader cannot offer — see
-                // `from_readers::drive_compaction_stream`). This path-based
-                // producer still owns its freshly-opened `reader` exclusively, so
-                // passing `scan_cancel` by reference here is behaviourally
-                // identical to the old `set_scan_cancel` + field-read.
-                //
-                // Pass the schema so the parser uses the real clustering column
-                // names; the header-inferred fallback uses generic names like
-                // "clustering_key", which would defeat extract_clustering_key.
-                // `drive_compaction_stream` is the SAME streaming helper the
-                // shared-reader producer thread uses (issue #2346) — the
-                // conversion/backpressure/cancellation-by-variant semantics are
-                // defined in exactly one place.
-                from_readers::drive_compaction_stream(
-                    &reader,
-                    run_index,
-                    &schema,
-                    &scan_cancel,
-                    &sender,
-                    sent_count.as_ref(),
-                )
-                .await
-            })
-        })();
-
-        if let Err(e) = stream_result {
-            // Forward the error (preserving `Cancelled` distinctly, issue #2264);
-            // ignore send failure (consumer may have dropped).
-            let _ = error_sender.send(Err(MergeProducerError::from(e)));
-        }
-        // Channel closed naturally when sender is dropped here.
-    }
-
-    /// Convert one streamed `(RowKey, Value, timestamp)` source entry into a
-    /// [`MergeEntry`] for run `run_index` (issue #827).
-    ///
-    /// Factored out of the producer loop so the streaming emit callback can call
-    /// it inline. Populates the clustering key from the decoded cells so wide-row
-    /// (clustering) partitions reconcile per `(pk, ck)` instead of collapsing
-    /// into one row.
-    fn build_merge_entry(
-        run_index: usize,
-        compaction_row: crate::storage::sstable::reader::compaction_row::CompactionRow,
-        schema: &TableSchema,
-    ) -> Result<MergeEntry> {
-        use crate::storage::sstable::reader::compaction_row::CompactionRowData;
-
-        let crate::storage::sstable::reader::compaction_row::CompactionRow {
-            key,
-            row_timestamp,
-            row_data,
-        } = compaction_row;
-        let decorated_key = DecoratedKey::from_key_bytes(key.as_bytes().to_vec())?;
-
-        // Issue #1072: a partition-level tombstone surfaces as a self-contained
-        // carrier `MergeEntry` (empty live row + `partition_deletion`, no
-        // clustering). `merge_partition_rows` extracts these, applies the MAX
-        // partition floor across sources to shadow older cells/rows/ranges, and
-        // re-emits the surviving partition tombstone. Handled BEFORE the
-        // `RangeMarker` block so the partition floor is the outermost shadow.
-        if let CompactionRowData::PartitionDelete {
-            deletion_time,
-            local_deletion_time,
-        } = row_data
-        {
-            // Carry the deletion in `RowData::Tombstone` so the carrier never
-            // surfaces as a phantom live row to consumers iterating the merge step
-            // stream; `partition_deletion` marks it as the partition-level carrier
-            // and `merge_entry_to_mutation` lifts it onto the partition header
-            // (emitting NO clustering-row `DeleteRow`).
-            return Ok(MergeEntry::new(
-                run_index,
-                decorated_key,
-                None,
-                // Use markedForDeleteAt as the entry timestamp so stats baselines
-                // see a real value (not 0) for the carrier.
-                deletion_time,
-                RowData::Tombstone {
-                    deletion_time,
-                    local_deletion_time,
-                },
-            )
-            .with_partition_deletion((deletion_time, local_deletion_time)));
-        }
-
-        // Issue #933: a range-tombstone marker surfaces as a self-contained
-        // carrier `MergeEntry` (empty live row + `range_deletion`), routed into the
-        // partition's `None` clustering bucket so it is never collapsed with data
-        // rows. `merge_partition_rows` extracts these to shadow covered cells and
-        // re-emit the surviving marker to the output SSTable.
-        if let CompactionRowData::RangeMarker {
-            start,
-            end,
-            deletion_time,
-            local_deletion_time,
-        } = row_data
-        {
-            let range = RangeTombstone {
-                start: Self::compaction_bound_to_mutation(start),
-                end: Self::compaction_bound_to_mutation(end),
-                deletion_time,
-                local_deletion_time,
-            };
-            return Ok(MergeEntry::new(
-                run_index,
-                decorated_key,
-                None,
-                // Use the deletion time as the entry timestamp so the writer's
-                // stats baselines see a real value (not 0) for the carrier.
-                deletion_time,
-                RowData::Live { cells: Vec::new() },
-            )
-            .with_range_deletion(range));
-        }
-        // #912: derive the clustering identity from the per-element compaction row
-        // BEFORE collapsing it to `RowData`. A row tombstone carries no cells in
-        // `RowData::Tombstone`, so its clustering key must come from the
-        // tombstone's captured clustering prefix; otherwise it would collapse into
-        // the partition's `None` bucket and mis-reconcile against the static row
-        // and against other clustering-row tombstones.
-        let clustering_key = Self::extract_clustering_key_from_compaction(&row_data, schema);
-        let (row_data, complex_deletions, row_deletion, row_liveness) =
-            Self::compaction_row_data_to_row_data(row_data, row_timestamp);
-        let entry = MergeEntry::new(
-            run_index,
-            decorated_key,
-            clustering_key,
-            row_timestamp,
-            row_data,
-        )
-        // Issue #2374/#2789: carry the row-marker liveness for the read path.
-        .with_row_liveness(row_liveness);
-        let entry = if complex_deletions.is_empty() {
-            entry
-        } else {
-            entry.with_complex_deletions(complex_deletions)
-        };
-        // Issue #932: carry the coexisting row deletion so reconciliation keeps it
-        // alongside the surviving live cells (preventing resurrection of older
-        // cells in non-compacted SSTables).
-        let entry = match row_deletion {
-            Some((deletion_time, ldt)) => entry.with_row_deletion(deletion_time, ldt),
-            None => entry,
-        };
-        Ok(entry)
-    }
-
-    /// Convert a reader-native [`CompactionBound`] into a write-engine
-    /// [`ClusteringBound`] (issue #933).
-    ///
-    /// [`CompactionBound`]: crate::storage::sstable::reader::compaction_row::CompactionBound
-    /// [`ClusteringBound`]: crate::storage::write_engine::mutation::ClusteringBound
-    fn compaction_bound_to_mutation(
-        bound: crate::storage::sstable::reader::compaction_row::CompactionBound,
-    ) -> crate::storage::write_engine::mutation::ClusteringBound {
-        use crate::storage::sstable::reader::compaction_row::CompactionBound;
-        use crate::storage::write_engine::mutation::ClusteringBound;
-        match bound {
-            CompactionBound::Inclusive(cols) => {
-                ClusteringBound::Inclusive(ClusteringKey { columns: cols })
-            }
-            CompactionBound::Exclusive(cols) => {
-                ClusteringBound::Exclusive(ClusteringKey { columns: cols })
-            }
-            CompactionBound::Bottom => ClusteringBound::Bottom,
-            CompactionBound::Top => ClusteringBound::Top,
-        }
-    }
-
-    /// Extract a `ClusteringKey` from the row's live cells using the schema.
-    ///
-    /// For each clustering column declared in the schema (in position order),
-    /// look for a cell with that column name in the decoded `RowData::Live`
-    /// cells.  If all clustering columns are found, return `Some(ClusteringKey)`;
-    /// otherwise (including for tombstone entries that have no cells) return
-    /// `None`.
-    ///
-    /// The clustering columns are intentionally left inside the cells so the
-    /// downstream read-back path can still find them.
-    fn extract_clustering_key_from_compaction(
-        row_data: &crate::storage::sstable::reader::compaction_row::CompactionRowData,
-        schema: &TableSchema,
-    ) -> Option<ClusteringKey> {
-        use crate::storage::sstable::reader::compaction_row::CompactionRowData;
-
-        if schema.clustering_keys.is_empty() {
-            return None;
-        }
-
-        match row_data {
-            // A live row surfaces its clustering columns as simple cells (#229).
-            CompactionRowData::Live { simple, .. } => {
-                let mut ck_columns: Vec<(String, Value)> =
-                    Vec::with_capacity(schema.clustering_keys.len());
-                for ck_col in &schema.clustering_keys {
-                    // Any missing clustering column ⇒ the WHOLE key is discarded (the `?`
-                    // returns `None`, treating the row as unclustered).
-                    let pair = simple
-                        .iter()
-                        .find(|c| c.column == ck_col.name)
-                        .map(|c| (ck_col.name.clone(), c.value.clone()))?;
-                    ck_columns.push(pair);
-                }
-                Some(ClusteringKey {
-                    columns: ck_columns,
-                })
-            }
-            // #912: a row tombstone carries its own clustering prefix so it
-            // reconciles in its own bucket instead of collapsing into `None`.
-            // An empty `clustering` (unclustered table / partial prefix) keeps the
-            // pre-#912 `None`-bucket behavior.
-            CompactionRowData::Tombstone { clustering, .. } => {
-                if clustering.is_empty() {
-                    return None;
-                }
-                // Reorder defensively into schema order; bail to `None` if any
-                // declared clustering column is absent.
-                let mut ck_columns: Vec<(String, Value)> =
-                    Vec::with_capacity(schema.clustering_keys.len());
-                for ck_col in &schema.clustering_keys {
-                    // Any missing clustering column ⇒ the WHOLE key is discarded (the `?`
-                    // returns `None`, treating the row as unclustered).
-                    let pair = clustering
-                        .iter()
-                        .find(|(name, _)| name == &ck_col.name)
-                        .map(|(name, v)| (name.clone(), v.clone()))?;
-                    ck_columns.push(pair);
-                }
-                Some(ClusteringKey {
-                    columns: ck_columns,
-                })
-            }
-            // Issue #933: a range marker is handled by `build_merge_entry` before
-            // this point and routed into the `None` clustering bucket.
-            CompactionRowData::RangeMarker { .. } => None,
-            // Issue #1072: a partition tombstone is handled by `build_merge_entry`
-            // before this point and routed into the `None` clustering bucket.
-            CompactionRowData::PartitionDelete { .. } => None,
-        }
-    }
-
-    /// Convert a [`CompactionRowData`] into the merge `RowData` plus the row's
-    /// complex-deletion markers (epic #899, Phase C — the behavioral flip).
-    ///
-    /// Simple columns become one [`CellData`] each (cell-own ts/ttl/ldt). A
-    /// complex (non-frozen collection / UDT) column is NO LONGER collapsed to a
-    /// single whole-column value: each [`ComplexElement`] becomes its OWN
-    /// per-element [`CellData`] carrying the element's authoritative `cell_path`,
-    /// per-element `timestamp`, `ttl`, `local_deletion_time`, `is_deleted`, and
-    /// on-disk `has_empty_value`. `reconcile_cluster` keys winners on
-    /// `(column, cell_path)`, so disjoint elements of the same column written
-    /// across SSTables all survive (Cassandra `Cells#reconcile`), and the
-    /// merge→mutation step emits a [`CellOperation::WriteComplexElement`] per
-    /// element — preserving the per-element on-disk layout byte-for-byte (epic
-    /// #899 north star). The reader's whole-collection `collapsed_value` is no
-    /// longer threaded to the writer (the per-element path is now authoritative);
-    /// it remains on the reader contract for user-facing reads.
-    ///
-    /// The real per-column complex deletion (`markedForDeleteAt` +
-    /// `localDeletionTime`) is surfaced on `complex_deletions` so the writer
-    /// emits a REAL deletion marker (replacing the LIVE sentinel). `reconcile_cluster`
-    /// reduces these to the strict-superseding (max-mfda) deletion per column NAME
-    /// and SHADOWS covered elements (ts <= mfda) before purge (issue #887). gc_grace
-    /// purging of the surviving marker remains future work (#845).
-    ///
-    /// [`ComplexElement`]: crate::storage::sstable::reader::compaction_row::ComplexElement
-    /// [`CellOperation::WriteComplexElement`]: crate::storage::write_engine::mutation::CellOperation::WriteComplexElement
-    #[allow(clippy::type_complexity)]
-    fn compaction_row_data_to_row_data(
-        row_data: crate::storage::sstable::reader::compaction_row::CompactionRowData,
-        _row_timestamp: i64,
-    ) -> (
-        RowData,
-        Vec<ComplexDeletion>,
-        Option<(i64, i32)>,
-        crate::storage::sstable::reader::compaction_row::RowLiveness,
-    ) {
-        use crate::storage::sstable::reader::compaction_row::{CompactionRowData, RowLiveness};
-
-        match row_data {
-            CompactionRowData::Tombstone {
-                deletion_time,
-                local_deletion_time,
-                // Clustering identity is consumed by
-                // `extract_clustering_key_from_compaction` before this conversion
-                // (#912); `RowData::Tombstone` itself carries only the timestamps.
-                clustering: _,
-            } => (
-                RowData::Tombstone {
-                    deletion_time,
-                    local_deletion_time,
-                },
-                Vec::new(),
-                None,
-                RowLiveness::default(),
-            ),
-            CompactionRowData::Live {
-                simple,
-                complex,
-                // Issue #932: a coexisting row deletion surfaces here so the
-                // merge entry carries it alongside the live cells.
-                row_deletion,
-                // Issue #2374/#2789: primary-key liveness carried for the read path.
-                row_liveness,
-            } => {
-                let element_count: usize = complex.iter().map(|c| c.elements.len()).sum();
-                let mut cells = Vec::with_capacity(simple.len() + element_count);
-                let mut complex_deletions = Vec::new();
-
-                for sc in simple {
-                    cells.push(CellData {
-                        column: sc.column,
-                        value: sc.value,
-                        timestamp: sc.timestamp,
-                        ttl: sc.ttl,
-                        cell_path: None,
-                        local_deletion_time: sc.local_deletion_time,
-                        is_complex_element: false,
-                        is_deleted: false,
-                        has_empty_value: false,
-                    });
-                }
-
-                for col in complex {
-                    // Surface the REAL complex deletion (replacing the LIVE
-                    // sentinel) so the writer emits a genuine marker (Phase C).
-                    if let Some((marked_for_delete_at, ldt)) = col.complex_deletion {
-                        complex_deletions.push(ComplexDeletion {
-                            column: col.column.clone(),
-                            marked_for_delete_at,
-                            local_deletion_time: ldt,
-                        });
-                    }
-
-                    // Per-element emit: one CellData per ComplexElement, keyed by
-                    // its authoritative cell_path, carrying per-element write
-                    // metadata verbatim. An element tombstone is represented with
-                    // `is_deleted = true` and a `Value::Tombstone(CellTombstone)`
-                    // value so the per-cell reconcile tie-break (tombstone beats
-                    // live at equal ts) still applies; a live element keeps its
-                    // decoded value (the reader stores the SET member decoded from
-                    // the path, but `has_empty_value` records the on-disk emptiness
-                    // so the writer reproduces it byte-for-byte).
-                    for elem in col.elements {
-                        let value = if elem.is_deleted {
-                            Value::Tombstone(Box::new(crate::types::TombstoneInfo {
-                                deletion_time: elem.timestamp,
-                                tombstone_type: crate::types::TombstoneType::CellTombstone,
-                                // Element's on-disk localDeletionTime (GC clock,
-                                // seconds); `0` when not surfaced (#873).
-                                local_deletion_time: elem.local_deletion_time.unwrap_or(0) as i64,
-                                ttl: None,
-                                range_start: None,
-                                range_end: None,
-                            }))
-                        } else {
-                            elem.value.unwrap_or(Value::Null)
-                        };
-                        cells.push(CellData {
-                            column: col.column.clone(),
-                            value,
-                            timestamp: elem.timestamp,
-                            ttl: elem.ttl,
-                            cell_path: Some(elem.cell_path),
-                            local_deletion_time: elem.local_deletion_time,
-                            is_complex_element: true,
-                            is_deleted: elem.is_deleted,
-                            has_empty_value: elem.has_empty_value,
-                        });
-                    }
-                }
-
-                (
-                    RowData::Live { cells },
-                    complex_deletions,
-                    row_deletion,
-                    row_liveness,
-                )
-            }
-            // Issue #933: range markers are intercepted in `build_merge_entry`
-            // (carrier entry with `range_deletion`); they never reach this
-            // conversion. Map defensively to an empty live row.
-            CompactionRowData::RangeMarker { .. } => (
-                RowData::Live { cells: Vec::new() },
-                Vec::new(),
-                None,
-                RowLiveness::default(),
-            ),
-            // Issue #1072: partition tombstones are intercepted in
-            // `build_merge_entry` (carrier entry with `partition_deletion`); they
-            // never reach this conversion. Map defensively to an empty live row.
-            CompactionRowData::PartitionDelete { .. } => (
-                RowData::Live { cells: Vec::new() },
-                Vec::new(),
-                None,
-                RowLiveness::default(),
-            ),
-        }
-    }
-
-    /// Convert a reader Value to RowData.
-    ///
-    /// `row_timestamp` is the per-row timestamp decoded from the on-disk row
-    /// header (see [`SSTableReader::iterate_all_partitions_for_compaction`]). The
-    /// reader does not surface per-cell timestamps for live cells, so each live
-    /// cell inherits the row timestamp. This is required for per-cell reconcile
-    /// and row-tombstone shadowing to compare cell timestamps correctly
-    /// (Issue #533) — without it live cells would default to 0 and be wrongly
-    /// shadowed by any row tombstone.
-    ///
-    /// Issue #505: `Value::Tombstone(RowTombstone)` is now correctly emitted by
-    /// the V5CompressedLegacy parser for deleted rows, and
-    /// `Value::Tombstone(CellTombstone)` appears inside `Value::Map` entries for
-    /// deleted cells.  Both are surfaced here so the merger can apply shadowing
-    /// semantics.  A cell tombstone keeps its own `deletion_time` so equal-ts
-    /// reconcile still resolves it correctly.
-    ///
-    /// Epic #899: superseded on the production path by
-    /// [`Self::compaction_row_data_to_row_data`] (the reader now surfaces
-    /// per-element [`CompactionRow`]s). Retained for the legacy-collapse merge
-    /// tests that assert the old whole-column collapse behavior.
-    #[cfg(test)]
-    fn value_to_row_data(value: &crate::types::Value, row_timestamp: i64) -> Result<RowData> {
-        match value {
-            crate::types::Value::Tombstone(info) => Ok(RowData::Tombstone {
-                deletion_time: info.deletion_time,
-                // #873: TombstoneInfo now carries the source localDeletionTime
-                // (GC clock, seconds). `RowData::Tombstone` stores it as i32
-                // (the on-disk width); the value fits an i32 in practice
-                // (epoch-seconds, incl. far-future negative-i32 bit patterns).
-                // NOTE: this legacy adapter is `#[cfg(test)]` only — the
-                // production compaction tombstone path is
-                // `compaction_row_data_to_row_data`, which already threads the
-                // LDT from `CompactionRowData::Tombstone`.
-                local_deletion_time: info.local_deletion_time as i32,
-            }),
-            crate::types::Value::Map(map_entries) => {
-                let mut cells = Vec::with_capacity(map_entries.len());
-                for (key, val) in map_entries {
-                    let column = match key {
-                        crate::types::Value::Text(s) => String::from_utf8_lossy(s).into_owned(),
-                        other => format!("{:?}", other),
-                    };
-                    // Cell tombstones carry their own deletion_time (Issue #505);
-                    // live cells inherit the row timestamp (Issue #533) so per-cell
-                    // shadowing and LWW order them against row tombstones correctly.
-                    let cell_ts = match val {
-                        crate::types::Value::Tombstone(info) => info.deletion_time,
-                        _ => row_timestamp,
-                    };
-                    cells.push(CellData {
-                        column,
-                        value: val.clone(),
-                        timestamp: cell_ts,
-                        // ttl / local_deletion_time / cell_path are threaded for
-                        // the followup behaviors (#844, #848) but the reader's
-                        // `(RowKey, Value, ts)` compaction stream does not yet
-                        // surface per-cell ttl, the cell's local-deletion-time,
-                        // or a complex-column cell-path (the map key here is the
-                        // top-level column name, not a collection element path),
-                        // so they stay `None`. Populating them is part of #844 /
-                        // #848 once the reader is extended (issue #886 plumbing).
-                        ttl: None,
-                        cell_path: None,
-                        local_deletion_time: None,
-                        is_complex_element: false,
-                        is_deleted: false,
-                        has_empty_value: false,
-                    });
-                }
-                Ok(RowData::Live { cells })
-            }
-            // Single value or other formats - wrap as a single cell
-            other => Ok(RowData::Live {
-                cells: vec![CellData {
-                    column: "value".to_string(),
-                    value: other.clone(),
-                    timestamp: row_timestamp,
-                    // Not surfaced by the reader's compaction stream yet; see the
-                    // note on the map-entry path above (issue #886 plumbing).
-                    ttl: None,
-                    cell_path: None,
-                    local_deletion_time: None,
-                    is_complex_element: false,
-                    is_deleted: false,
-                    has_empty_value: false,
-                }],
-            }),
-        }
-    }
-}
-
-#[cfg(feature = "write-support")]
-impl SSTableRowIterator for SSTableRowIteratorAdapter {
-    fn next(&mut self) -> Option<Result<MergeEntry>> {
-        use std::sync::mpsc::RecvTimeoutError;
-        // The receiver is only `None` after `Drop` has run — never during a live
-        // merge — so treat a missing receiver as end-of-stream.
-        let receiver = self.receiver.as_ref()?;
-        // Cancel-aware blocking recv (issue #2361): a plain `recv()` blocks
-        // indefinitely inside `KWayMerger::step`, so a `do_get` cancellation that
-        // lands while the consumer is waiting for the next producer entry was NOT
-        // observed until an entry (or channel close) arrived. Poll `scan_cancel`
-        // at a bounded interval so `drive_merge` sees `Error::Cancelled` promptly
-        // even while blocked here.
-        loop {
-            if self.scan_cancel.is_cancelled() {
-                return Some(Err(Error::Cancelled));
-            }
-            match receiver.recv_timeout(RECV_CANCEL_POLL) {
-                Ok(Ok(entry)) => {
-                    // Issue #2419 (WS2): this DATA entry just left the bounded
-                    // egress channel — decrement the live occupancy gauge,
-                    // balancing the `channel_depth::sent()` at its send site
-                    // (`from_readers::forward_row`). `received_count` is this
-                    // adapter's OWN mirror of that decrement (roborev job 1733),
-                    // read by `Drop` post-join to compute the exact reconcile
-                    // residual — see the field doc.
-                    channel_depth::received();
-                    self.received_count += 1;
-                    // Issue #2096: one merge entry decoded from `Data.db` by THIS
-                    // adapter-driven run — a full scan, compaction, or (via the
-                    // fail-safe `SinglePartitionFilterRun`) a point read all share
-                    // this increment site, so `merge_run_entries_decoded` counts
-                    // entries for any of them, not point reads alone. See
-                    // `work_counters::merge_run_entries_decoded`'s doc for the
-                    // process-global caveat when using it as a delta assertion.
-                    crate::storage::sstable::work_counters::add_merge_run_entry_decoded();
-                    return Some(Ok(entry));
-                }
-                // Issue #2264: reconstruct `Error::Cancelled` distinctly so a
-                // cancelled scan is never confused with a genuine I/O/corruption
-                // error at the merge/producer boundary — `drive_merge` matches on
-                // the variant, not on a side-channel flag.
-                Ok(Err(MergeProducerError::Cancelled)) => return Some(Err(Error::Cancelled)),
-                Ok(Err(MergeProducerError::Other(msg))) => {
-                    return Some(Err(Error::Storage(format!(
-                        "streaming merge producer error: {}",
-                        msg
-                    ))))
-                }
-                // No entry yet: re-poll the cancel flag and keep waiting.
-                Err(RecvTimeoutError::Timeout) => continue,
-                // Channel closed — producer finished normally.
-                Err(RecvTimeoutError::Disconnected) => return None,
-            }
-        }
-    }
-
-    /// Issue #2765 wiring hook: the exact adaptive capacity this adapter's egress
-    /// `sync_channel` was constructed with (the merge-scoped snapshot the
-    /// constructor threaded in).
-    #[cfg(test)]
-    fn egress_channel_capacity(&self) -> Option<usize> {
-        Some(self.egress_channel_capacity)
-    }
-}
-
-/// Cancel-aware teardown (issue #2361): trip the scan token, close the channel,
-/// then JOIN the producer thread — so a dropped merger (LIMIT satisfied, client
-/// disconnect, error, panic) does not leave a detached producer streaming a
-/// multi-million-partition scan in the background.
-#[cfg(feature = "write-support")]
-impl Drop for SSTableRowIteratorAdapter {
-    fn drop(&mut self) {
-        // 1. Trip the cooperative token so a producer that is mid-scan (polling
-        //    `scan_cancel`, NOT blocked on a channel send) abandons its walk.
-        self.scan_cancel.cancel();
-        // 2. Drop the receiver so a producer BLOCKED on a full `SyncSender::send`
-        //    wakes immediately (send returns `Err`, and `drive_compaction_stream`
-        //    maps that to `ControlFlow::Break`). Without this the join below could
-        //    block until the channel drained. Dropping it here (before the join)
-        //    rather than relying on field-drop order (which runs AFTER this `Drop`)
-        //    is what makes the join bounded.
-        //
-        //    Issue #2419 roborev job 1733: do NOT try to drain-and-decrement the
-        //    egress-depth gauge HERE. A prior version looped `try_recv` until
-        //    `Empty` and decremented per drained entry before dropping the
-        //    receiver — but the producer runs on another OS thread, so a send can
-        //    slip in and succeed (incrementing the shared gauge) in the window
-        //    between that loop's last `Empty` and this `drop`, with no entry ever
-        //    physically drained to balance it: a PERMANENT, monotonic upward leak
-        //    of `cqlite.merge.egress_channel_depth` across every
-        //    cancelled/disconnected scan on a long-running server. The gauge is
-        //    reconciled instead in step 4, AFTER the join, using this adapter's
-        //    own sent/received delta — which only becomes authoritative once the
-        //    producer thread has provably stopped sending.
-        drop(self.receiver.take());
-        // 3. Join the producer — bounded, because after (1)+(2) the producer
-        //    cannot block indefinitely: it either observes the cancel in its scan
-        //    loop or its next send fails. A failed join (producer panicked) is
-        //    ignored; the panic was already surfaced through the error channel.
-        if let Some(handle) = self.producer.take() {
-            let _ = handle.join();
-        }
-        // 4. Issue #2419 roborev job 1733: reconcile the shared egress-depth
-        //    gauge AFTER the producer thread has definitively exited — `join`
-        //    returning guarantees `self.sent_count` can never increase again, so
-        //    this read is race-free (unlike reading it, or draining the channel,
-        //    before the join). Any DATA entries this adapter's producer
-        //    successfully sent but this consumer never received — abandoned when
-        //    the channel was torn down while entries were still buffered —
-        //    are exactly `sent_count - received_count`; reconcile the shared
-        //    gauge by that residual in ONE atomic op so a cancelled/disconnected
-        //    scan returns `cqlite.merge.egress_channel_depth` to baseline instead
-        //    of drifting upward.
-        let residual =
-            self.sent_count.load(std::sync::atomic::Ordering::SeqCst) - self.received_count;
-        channel_depth::reconcile_residual(residual);
-    }
-}
 
 /// K-way merger for combining multiple SSTables
 ///
